the TrueType font now used in menus....
.... it acted crazy..... i thought i was applying it only in one area,
and yet the more i ran the game, the more of the ingame text it was
switching over to using the new font.   like all the buttons at first
remained the same thin default text, but the longer i ran the game the
more they would all the sudden change themselves to the new text.