Roguelike, particles
the rocket explosion now looks better, and the roguelike now generates
proper 2d(it got messed up when I started switching to 2d). Also noticed
a bug with the avatar class(the energy is ticked with move() and not
update() which sometimes causes it to be restored faster), but didn't
fix it since we'll be getting a new avatar system anyways.