Roguelike is now 3d
A single level generates properly - but proceeding to dungeons will not
be hard, since they're already mostly coded, they just need calls to
Build3D() and setting the locations of every level, together with making
holes in ceilings/floors to interconnect them. Probably will need a
staircase asset to be put there. Also roguelike will need proper
textures and lights. And a gamemode with AI. You can move the cam with
w/s + mouse motion, press space to toggle cam trail(used it because it
was hard to navigate in the dungeons).