removed spamtula again...
...wasn't much better the 2nd time around.  8)  i was annoyed to see and
hear other players hitting their mouse buttons.  all over again.  which
is why i deleted it in the 1st place.

i would allow about %25 of the original spamtula messages to be reworked
as obituary messages if you wanted to do that.   however i'm not certain
i want to keep a spatula as a melee weapon